Move configurator.py to _internal
