+++ certbot-nginx/setup.py
@@ -79,7 +79,7 @@ setup(
     },
     entry_points={
         'certbot.plugins': [
-            'nginx = certbot_nginx.configurator:NginxConfigurator',
+            'nginx = certbot_nginx._internal.configurator:NginxConfigurator',
         ],
     },
     test_suite='certbot_nginx',
